[DEV] Refactor and Feature Enhancements for Dashboard and Sidebar (#25)
* chore(config): Add melancong protocols

* refactor(dashboard): Refcator home route to dashboard

* feat(card): Add card destinations

* feat(sidebar): Add sidebar as partials of dashboard page

* refactor(Logo): Refactor logo sidebar as partials

* refactor(profile): Move sidebar profile inside dashboard route

* chore(pages): Remove unused route

* refactor(route): Move route from init state to dashboard's route

* chore(placeholder): Add argument to ctx's getcontext function

* refactor(navigation): Modify route based on their purpose
