+++ app/(route)/dashboard/partials/DashboardDestination.tsx
@@ -14,7 +14,7 @@ interface Props {
   filterSubmitted: boolean
 }
 
-const DashboardUser: NextPage<Props> = ({
+const DashboardDestination: NextPage<Props> = ({
   destinations,
   searchQuery,
   handleSearch,
@@ -48,7 +48,7 @@ const DashboardUser: NextPage<Props> = ({
 
   return (
     <>
-      <div className='flex flex-1'>
+      <div className='size-full'>
         <div className='flex h-full w-full flex-1 flex-col gap-2 rounded-tl-2xl border border-neutral-200 bg-white p-2 dark:border-neutral-700 dark:bg-neutral-900 md:p-10'>
           <div className='py-4'>
             <PlaceholdersAndVanishInput
@@ -93,4 +93,4 @@ const DashboardUser: NextPage<Props> = ({
   )
 }
 
-export default DashboardUser
+export default DashboardDestination
