+++ app/(route)/dashboard/partials/Logo.tsx
