+++ app/(route)/dashboard/partials/LogoIcon.tsx
