+++ components/ui/input/placeholders-and-vanish-input.tsx
@@ -59,7 +59,7 @@ export function PlaceholdersAndVanishInput({
     if (!canvas) {
       return
     }
-    const ctx = canvas.getContext('2d')
+    const ctx = canvas.getContext('2d', { willReadFrequently: true })
     if (!ctx) {
       return
     }
