+++ data/index.tsx
@@ -192,21 +192,21 @@ export const destinationTypes = [
 export const links = [
   {
     label: 'Dashboard',
-    href: '#',
+    href: '/dashboard',
     icon: (
       <IconBrandTabler className='h-5 w-5 flex-shrink-0 text-neutral-700 dark:text-neutral-200' />
     )
   },
   {
     label: 'Profile',
-    href: '#',
+    href: '/dashboard/profile',
     icon: (
       <IconUserBolt className='h-5 w-5 flex-shrink-0 text-neutral-700 dark:text-neutral-200' />
     )
   },
   {
     label: 'Settings',
-    href: '#',
+    href: '/dashboard/settings',
     icon: (
       <IconSettings className='h-5 w-5 flex-shrink-0 text-neutral-700 dark:text-neutral-200' />
     )
@@ -220,7 +220,7 @@ export const links = [
   },
   {
     label: 'Logout',
-    href: '#',
+    href: '/auth/logout',
     icon: (
       <IconArrowLeft className='h-5 w-5 flex-shrink-0 text-neutral-700 dark:text-neutral-200' />
     )
